License under GPL v2 (same as Inkscape itself).
Note that I do not regard the software in this repo as a derivative
work of Inkscape as it dynamically accesses the publicly exposed
commandlines only and alternative SVG software or libraries could be
dropped in. However I have decided to license it in a similar manner
for the benefit of the community.